Add initial Travis-CI config
Travis-CI is a continuous integration service which is free for open
source projects.  This change adds an initial configuration for
this service.  Note that xvfb is used to handle the fact that Travis CI
is a headless environment and that one needs to use the `-a` flag to
`xvfb-run` in order that the `xvfb` server is automatically started,
which then allows the dist to build and test reliably on this
infrastructure.