Fix tklib detection
As mentioned in RT#125850, the check for tklib's existence can fail even
if the library has been installed.  This change tests for the existence
of the `cursor` tklib package which was part of the initial tklib import
commit (https://core.tcl.tk/tklib/info/ec23b04ac240ca3c) and still
exists in the library, hence this check should now be fairly robust and
backwards compatible over (hopefully) all versions of tklib.