Add exit button to puzzle demo
An exit button is handy so that the user can go from the main page back
to the main page after having played with the demo.  This change
includes an update to allow the script to run on its own without the
need to call the main `widget` demo program.